simple-host-bot/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.5
aiosignal==1.3.1
aiounittest==1.3.0
async-timeout==4.0.3
attrs==24.2.0
azure-core==1.32.0
Babel==2.9.1
botbuilder-core==4.16.2
botbuilder-dialogs==4.16.2
botbuilder-integration-aiohttp==4.16.2
botbuilder-schema==4.16.2
botframework-connector==4.16.2
botframework-streaming==4.16.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
datedelta==1.4
emoji==1.7.0
frozenlist==1.5.0
grapheme==0.6.0
idna==3.10
isodate==0.7.2
jsonpickle==1.4.2
msal==1.31.0
msrest==0.7.1
multidict==6.1.0
multipledispatch==1.0.0
oauthlib==3.2.2
propcache==0.2.0
pycparser==2.22
PyJWT==2.9.0
pytz==2024.2
recognizers-text==1.0.2a2
recognizers-text-choice==1.0.2a2
recognizers-text-date-time==1.0.2a2
recognizers-text-number==1.0.2a2
recognizers-text-number-with-unit==1.0.2a2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
typing_extensions==4.12.2
urllib3==2.2.3
yarl==1.17.1
